renamed constants2 to just constants.ts
